requirements.txt
gitpython
mistralai
